+++ R/env_filter.R
@@ -12,22 +12,40 @@
 #'
 #' @param predictors A data frame of predictors. Must include atleast one the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, bio17. Columns not included must then be included in the 'exclude' argument.
 #' @param taxa A vector of strings containing one or more taxa to generate predictions for.
+#' @param screen A boolean (TRUE/FALSE) indicating whether to use the `elements::envelope_filter` function to check whether taxa are within distibutional limits prior to applying the more computationally expensive `elements::predict_occ` or `elements::calc_distance` functions.
 #' @param method One of "svm", "mean", or "median".
 #' @param limit A string representing the niche width quantiles, one of "min_max", "q01_q99", "q05_q95", "q25_q75". Which if set assigns a probability of 0 to a set of predictors if one or more of those predictors are outside the stipulated quantile ranges. Only applied if pa = "Present". Optional.
 #' @param exclude Model variables to exclude from the distance calculation; passed to the 'holdout' argument of 'elements::predict_occ' if the 'method' argument is "svm", otherwise when the 'method' argument is set to "mean" or "median" those variables are removed from the distance calculation.
 #' @param threshold A probability threshold to use as a cut off in the environmental filter. Only applicable when 'method' = "svm".
-#' @param toptaxa The number of top taxa, as defined by their probability when 'method' = "svm" or euclidean distance when when 'method' = "mean" or "median".
+#' @param append A string, one of "all", "predictors", or "ids" representing which columns from the predictors data frame to return with the results.
 #'
 #' @returns A dataframe containing three columns: taxon_code, rank, and Present (if 'method' = "svm") or Distance (if 'method' = "mean" or "median").
 #' @export
 #'
 #' @examples
-#' elements::startup(); elements::env_filter(predictors = elements::ExampleScenarios[1,], taxa = elements::TaxonomicBackbone$taxon_code, method = "svm", threshold = 0.5)
-env_filter <- function(predictors = elements::ExampleScenarios[1,], taxa = elements::TaxonomicBackbone$taxon_code, method = "svm", limit = NULL, exclude = NULL, threshold = NULL, toptaxa = NULL){
+#' elements::startup(); elements::env_filter(predictors = elements::ExampleScenarios, taxa = elements::TaxonomicBackbone$taxon_code, method = "svm", threshold = 0.5)
+env_filter <- function(predictors, taxa = elements::TaxonomicBackbone$taxon_code, screen = TRUE, method = "svm", limit = "min_max", exclude = NULL, threshold = NULL, append = "ids"){
   
-  if(isFALSE(nrow(predictors) == 1)){
-    stop("Only one row of predictors may be used with `elements::env_filter_plot` use `elements::env_filter` instead.")
+  if(screen == TRUE){
+    
+    screened_data <- elements::envelope_filter(taxa = taxa,
+                                               predictors = predictors, 
+                                               vars = elements::VariableNames, 
+                                               limit = limit)
+    
+    inside_range <- subset(screened_data, within_limits == TRUE, select = -within_limits)
+    
+    outside_range <- subset(screened_data, within_limits == FALSE, select = -within_limits)
+    
+  } else {
+    
+    inside_range <- merge(predictors, data.frame("taxon_code" = taxa))
+    
+    outside_range <- inside_range[0,]
+    
   }
+  
+  taxa_results_outside_range <- outside_range
   
   if(method == "svm"){
     
@@ -37,94 +55,43 @@ env_filter <- function(predictors = elements::ExampleScenarios[1,], taxa = eleme
     }
     
     # Retrieve probabilities for all taxa and predictors
-    taxa_results <- elements::predict_occ(taxa_codes = taxa,
-                                          predictors = predictors,
-                                          limit = limit,
-                                          holdopt = exclude,
-                                          append_predictors = FALSE)
+    taxa_results_inside_range <- elements::predict_occ(taxa = NULL,
+                                                       predictors = inside_range,
+                                                       limit = NULL,
+                                                       holdopt = exclude,
+                                                       append = append)
     
-    taxa_results <- taxa_results[order(taxa_results$Present, decreasing = TRUE), ]
-    
-    if(!is.null(threshold)){
-      
-      taxa_results_filtered <- subset(taxa_results, Present >= threshold)[,c("taxon_code", "Present")]
-      
-    } else if(is.null(threshold) & !is.null(toptaxa)){
-      
-      taxa_results_filtered <- taxa_results[1:toptaxa, c("taxon_code", "Present")]
-      
-    } else if(is.null(threshold) & is.null(toptaxa)) {
-      
-      taxa_results_filtered <- taxa_results[,c("taxon_code", "Present")]
-      
-    }
+    if(nrow(outside_range) > 0){taxa_results_outside_range["Present"] <- 0}
     
   } else if(method %in% c("mean", "median")){
     
-    predictors_mat <- as.matrix(predictors[, setdiff(elements::VariableNames, exclude)])
+    taxa_results_inside_range <- elements::calc_distance(predictors = inside_range, 
+                                                         taxa = taxa, 
+                                                         vars = elements::VariableNames, 
+                                                         method = method,
+                                                         append = append)
     
-    optima_df <- subset(elements::NicheWidths, taxon_code %in% taxa, select = c("variable", method, "taxon_code"))
-    
-    if(!is.null(exclude)){
-      optima_df <- subset(elements::NicheWidths, !(variable %in% exclude))
-    }
-    
-    optima_mat <- stats::xtabs(get(method) ~ taxon_code + variable, data = optima_df)
-    
-    attr(optima_mat, "class") = NULL; attr(optima_mat, "call") = NULL; dimnames(optima_mat) = unname(dimnames(optima_mat))
-    
-    dist_from_optima <- sweep(optima_mat, 2, predictors_mat) ^ 2
-      
-    dist_from_optima <- scale(dist_from_optima, center = FALSE)
-    
-    taxa_results <- sqrt(rowSums(dist_from_optima))
-    
-    taxa_results <- data.frame("taxon_code" = names(taxa_results), "Distance" = unname(taxa_results))
-    
-    # Need to optimise this!
-    if(!is.null(limit)){
-      
-      for(taxon in taxa){
-        
-        nw <- elements::NicheWidths
-        nw_taxon <- subset(nw[nw[["taxon_code"]] == taxon, ], select = -taxon_code)
-        nw_taxon <- stats::setNames(data.frame(t(nw_taxon[,-1])), nw_taxon[[1]])
-        
-        lower <- nw_taxon[unlist(strsplit(limit, "_"))[1], ]
-        upper <- nw_taxon[unlist(strsplit(limit, "_"))[2], ]
-        
-        vars <- colnames(lower)
-        
-        for(var in vars){
-            taxa_results[taxa_results["taxon_code"] == taxon, "Distance"] <- ifelse(predictors_mat[, var] > upper[[var]], NA, taxa_results[taxa_results["taxon_code"] == taxon, "Distance"])
-            taxa_results[taxa_results["taxon_code"] == taxon, "Distance"] <- ifelse(predictors_mat[, var] < lower[[var]], NA, taxa_results[taxa_results["taxon_code"] == taxon, "Distance"])
-        }
-        
-      }
-      
-    }
-    
-    taxa_results <- taxa_results[order(taxa_results$Distance, decreasing = FALSE), ]
-    
-    if(!is.null(threshold)){
-      
-      stop("The threshold argument can only be used when method = \"svm\".")
-      
-    } else if(is.null(threshold) & !is.null(toptaxa)){
-      
-      taxa_results_filtered <- taxa_results[1:toptaxa, ]
-      
-    } else if(is.null(threshold) & is.null(toptaxa)) {
-      
-      taxa_results_filtered <- taxa_results
-      
-    }
+    if(nrow(outside_range) > 0){taxa_results_outside_range["distance"] <- Inf}
     
   }
   
-  rownames(taxa_results_filtered) <- NULL
+  taxa_results_outside_range <- taxa_results_outside_range[, colnames(taxa_results_inside_range)]
   
-  taxa_results_filtered["rank"] <- as.numeric(rownames(taxa_results_filtered))
+  taxa_results_all <- rbind(taxa_results_inside_range, taxa_results_outside_range)
+  
+  # Subset taxa
+  if(!is.null(threshold)){
+    
+    taxa_results_filtered <- subset(taxa_results_all, Present >= threshold)
+    
+  } else if(is.null(threshold)) {
+    
+    taxa_results_filtered <- taxa_results_all
+    
+  }
+    
+  # Sort
+  rownames(taxa_results_filtered) <- NULL
   
   return(taxa_results_filtered)