+++ R/predict_occ.R
@@ -1,7 +1,7 @@
 #' Generate predictions for the probability of occurrence for specified taxa
 #' 
 #' Generate predictions for the probability of occurrence (presence and/or absence) using a data frame of predictors for a one or more taxa specified in either:
-#' 1) the 'taxa_codes' argument, or
+#' 1) the 'taxa' argument, or
 #' 2) a column in the predictors data frame named 'taxon_code'
 #'
 #' A number of optional arguments provide additional control over the use of the models.
@@ -17,21 +17,21 @@
 #' This may be useful in instances where you wish to remove the influence of one or more variables
 #' on the model results.
 #' 
-#' If taxon codes are supplied in the 'taxa_codes' argument and there is a column in the predictors data frame named 'taxon_code', 
-#' the taxa present in the 'taxa_codes' argument will be used and results will be calculated using the entire predictors data frame
+#' If taxon codes are supplied in the 'taxa' argument and there is a column in the predictors data frame named 'taxon_code', 
+#' the taxa present in the 'taxa' argument will be used and results will be calculated using the entire predictors data frame
 #' and the 'taxon_code' column will be replaced.
 #'
 #' NOTE: to use this function you must first run `elements::startup()`
 #'
-#' @param taxa_codes A vector of strings containing one or more taxa to generate predictions for. Optional.
-#' @param predictors A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, bio17, and taxon
+#' @param taxa A vector of strings containing one or more taxa to generate predictions for. See `elements::TaxonomicBackbone`. Optional.
+#' @param predictors A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, bio17, and optionally taxon
 #' @param pa One of "Present", "Absent", or c("Present", "Absent").
 #' @param limit A string representing the niche width quantiles, one of "min_max", "q01_q99", "q05_q95", "q25_q75". Which if set assigns a probability of 0 to a set of predictors if one or more of those predictors are outside the stipulated quantile ranges. Only applied if pa = "Present". Optional.
 #' @param holdopt Hold one or more variables at their optimum values. NULL by default, else a vector of variable codes, e.g. c("SD", "GP").
 #' @param dp The number of decimal places to round the probability values to.
-#' @param append_predictors A boolean. If TRUE return the predictors data frame with the results in an additional column.
+#' @param append A string, one of "all", "predictors", or "ids" representing which columns from the predictors data frame to return with the results.
 #'
-#' @return A data frame containing the probability of occurrence (Present and/or Absent) for each taxon, if specified, appended to the predictorsS data.
+#' @return A data frame containing the probability of occurrence (Present and/or Absent) for each taxon.
 #' @export
 #' 
 #' @examples
@@ -39,17 +39,17 @@
 #' elements::startup()
 #' 
 #' # Generate predictions using a data frame containing taxon_codes in the 'taxon' column.
-#' elements::predict_occ(taxa_codes = NULL, predictors = elements::ExampleData2, pa = "Present", limit = NULL, holdopt = c("SD", "GP"), dp = 3, append_predictors = FALSE)
+#' elements::predict_occ(taxa = NULL, predictors = elements::ExampleData2, pa = "Present", limit = NULL, holdopt = c("SD", "GP"))
 #' 
-#' # Generate predictions for taxa specified in the 'taxa_codes' argument using a data frame containing only predictor variables.
-#' elements::predict_occ(taxa_codes = c("stellaria_graminea", "silene_flos-cuculi"), predictors = elements::ExampleData1, holdopt = c("SD", "GP"), pa = "Present", limit = NULL, dp = 3, append_predictors = TRUE)
+#' # Generate predictions for taxa specified in the 'taxa' argument using a data frame containing only predictor variables.
+#' elements::predict_occ(taxa = c("stellaria_graminea", "silene_flos-cuculi"), predictors = elements::ExampleData1, pa = "Present", limit = NULL, holdopt = c("SD", "GP"))
 #' 
-#' # Generate predictions for taxa specified in the 'taxa_codes' argument using a data frame containing both predictor variables and a 'taxon_code' column (which is ignored and overwritten).
-#' elements::predict_occ(taxa_codes = c("stellaria_graminea", "silene_flos-cuculi"), predictors = elements::ExampleData2, holdopt = c("SD", "GP"), pa = "Present", limit = NULL, dp = 3, append_predictors = TRUE)
+#' # Generate predictions for taxa specified in the 'taxa' argument using a data frame containing both predictor variables and a 'taxon_code' column (which is ignored and overwritten).
+#' elements::predict_occ(taxa = c("stellaria_graminea", "silene_flos-cuculi"), predictors = elements::ExampleData2, pa = "Present", limit = NULL, holdopt = c("SD", "GP"))
 #' 
 #' elements::shutdown() 
 #' }
-predict_occ <- function(taxa_codes, predictors, pa = "Present", limit = NULL, holdopt = NULL, dp = 3, append_predictors = TRUE){
+predict_occ <- function(taxa, predictors, pa = "Present", limit = NULL, holdopt = NULL, dp = 3, append = "ids"){
   
   # Check whether elements::startup() has been run and the Models filehashDB1 object is in the global environment
   if(isFALSE(exists(x = "Models", envir = elementsEnv))){
@@ -68,7 +68,7 @@ predict_occ <- function(taxa_codes, predictors, pa = "Present", limit = NULL, ho
     }
   }
   
-  if(is.null(taxa_codes) & "taxon_code" %in% colnames(predictors)){
+  if(is.null(taxa) & "taxon_code" %in% colnames(predictors)){
     
     # Split
     predictors_list <- split(predictors, predictors[["taxon_code"]])
@@ -83,19 +83,18 @@ predict_occ <- function(taxa_codes, predictors, pa = "Present", limit = NULL, ho
                                                                    limit = limit,
                                                                    holdopt = holdopt,
                                                                    dp = dp,
-                                                                   append_predictors = append_predictors)
-                             
+                                                                   append = append)
                              result[["taxon_code"]] <- taxon
-                             
+
                              return(result)
                              
                            }
     )
     
-  } else if(!is.null(taxa_codes)){
+  } else if(!is.null(taxa)){
     
     # Apply
-    results_list <- lapply(X = taxa_codes, 
+    results_list <- lapply(X = taxa, 
                            FUN = function(taxon){
                              
                              result <- elements::predict_occ_taxon(taxon = taxon,
@@ -104,7 +103,7 @@ predict_occ <- function(taxa_codes, predictors, pa = "Present", limit = NULL, ho
                                                                    limit = limit,
                                                                    holdopt = holdopt,
                                                                    dp = dp,
-                                                                   append_predictors = append_predictors)
+                                                                   append = append)
                              
                              result[["taxon_code"]] <- taxon
                              
@@ -113,9 +112,9 @@ predict_occ <- function(taxa_codes, predictors, pa = "Present", limit = NULL, ho
                            }
     )
     
-  } else if(is.null(taxa_codes) & !("taxon" %in% colnames(predictors))) {
+  } else if(is.null(taxa) & !("taxon" %in% colnames(predictors))) {
     
-    stop("The taxa to model must be specified in either: 1) the 'taxa_codes' argument, 2) or present in a column in the predictors data frame named 'taxon_code'")
+    stop("The taxa to model must be specified in either: 1) the 'taxa' argument, 2) or present in a column in the predictors data frame named 'taxon_code'")
     
   }