+++ tests/testthat/test-predict_occ_taxon.R
@@ -9,7 +9,7 @@ testthat::test_that("predict_occ_taxon works", {
   test_pa <- c("Present", "Absent")
   
   actual <- elements::predict_occ_taxon(taxon = test_taxon, predictors = test_predictors, 
-                                        pa = test_pa, limit = NULL, holdopt = NULL, dp = 3, append = TRUE)
+                                        pa = test_pa, limit = NULL, holdopt = NULL, dp = 3, append = "all")
   
   testthat::expect_equal(colnames(actual), c(colnames(test_predictors), test_pa))
   testthat::expect_equal(nrow(actual), nrow(test_predictors))
@@ -29,7 +29,7 @@ testthat::test_that("predict_occ_taxon works, holding GP and SD at optima", {
   test_pa <- c("Present", "Absent")
   
   actual <- elements::predict_occ_taxon(taxon = test_taxon, predictors = test_predictors, 
-                                        pa = test_pa, limit = NULL, holdopt = c("GP", "SD"), dp = 3, append = TRUE)
+                                        pa = test_pa, limit = NULL, holdopt = c("GP", "SD"), dp = 3, append = "all")
   
   testthat::expect_equal(unique(actual[["GP"]]), subset(elements::NicheWidths, 
                                                         variable == "GP" & taxon_code == test_taxon,